+++ src/middlewares/middlewares.js
@@ -3,12 +3,12 @@ import { getFriends, getName } from "rdf-query/rdf-query";
 import {
   loadEmailError,
   loadEmailSuccess,
-  loadFriendsSuccess
+  loadFriendsSuccess,
 } from "../store/actions/UserActions";
 import { getRoutesFromPod } from "../solid/routes";
 import {
   loadRoutesSuccess,
-  loadRoutesError
+  loadRoutesError,
 } from "../store/actions/RouteActions";
 
 export const myLogger = (store) => (next) => (action) => {
@@ -23,6 +23,7 @@ export const asyncRouteFetch = (store) => (next) => (action) => {
       if (webId) {
         getRoutesFromPod(webId)
           .then((routes) => {
+            console.log(routes);
             store.dispatch(loadRoutesSuccess(routes));
           })
           .catch((error) => {
@@ -63,7 +64,7 @@ export const asyncProfileFetch = (store) => (next) => (action) => {
               for (let i = 0; i < friendsNames.length; i++) {
                 let friend = {
                   name: friendsNames[i],
-                  uri: friendsUri[i]
+                  uri: friendsUri[i],
                 };
                 friendsObjects.push(friend);
               }
